Move build version number to version package
