+++ .goreleaser.yaml
@@ -18,7 +18,7 @@ builds:
     binary: bin/bk
     main: ./cmd/bk
     ldflags:
-      - -s -w -X github.com/buildkite/cli/v3/internal/build.Version={{.Version}}
+      - -s -w -X github.com/buildkite/cli/v3/internal/version.Version={{.Version}}
 
   - id: linux
     goos: [linux]
@@ -28,7 +28,7 @@ builds:
     binary: bin/bk
     main: ./cmd/bk
     ldflags:
-      - -s -w -X github.com/buildkite/cli/v3/internal/build.Version={{.Version}}
+      - -s -w -X github.com/buildkite/cli/v3/internal/version.Version={{.Version}}
 
   - id: windows
     goos: [windows]
@@ -36,7 +36,7 @@ builds:
     binary: bin/bk
     main: ./cmd/bk
     ldflags:
-      - -s -w -X github.com/buildkite/cli/v3/internal/build.Version={{.Version}}
+      - -s -w -X github.com/buildkite/cli/v3/internal/version.Version={{.Version}}
 
 archives:
   - id: macos-archive
