+++ internal/version/version.go
@@ -1,4 +1,4 @@
-package build
+package version
 
 // Version is dynamically set at build time through ldflags
 var Version = "DEV"
